backend/pyproject.toml
[project]
name = "gst-recon-backend"
version = "1.0.0"
description = "Knowledge Graph-based GST Reconciliation Engine"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.32.0",
    "neo4j==5.26.0",
    "openai==1.58.0",
    "google-generativeai==0.8.0",
    "ollama>=0.4.0",
    "scikit-learn==1.6.0",
    "networkx==3.4.0",
    "pandas==2.2.0",
    "python-multipart==0.0.18",
    "pydantic==2.10.0",
    "pydantic-settings==2.7.0",
    "python-dotenv==1.0.1",
    "httpx>=0.27.0",
    "jinja2==3.1.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]
